fixes #4
