+++ .gitignore
@@ -5,3 +5,4 @@
 bin/
 build/
 run/
+.idea/
