+++ gradle.properties
@@ -1,4 +1,4 @@
 org.gradle.jvmargs=-Xmx3G
 minecraft_version=1.12.2
 short_version=1.12
-mod_version=1.0.2
+mod_version=1.0.3
