+++ src/main/java/customblockfabricator/proxy/CommonProxy.java
@@ -10,4 +10,8 @@ public class CommonProxy {
   public void registerRenderers() {
     
   }
+
+  public void registerItemAndBlockColors() {
+
+  }
 }
